move format_analyzed_snippets to server.utils
Fixes: #192

Addressing:
Traceback (most recent call last):
  File "/usr/bin/logdetective", line 5, in <module>
    from logdetective.logdetective import main
  File "/usr/lib/python3.13/site-packages/logdetective/logdetective.py", line 6, in <module>
    from logdetective.utils import (
    ...<6 lines>...
    )
  File "/usr/lib/python3.13/site-packages/logdetective/utils.py", line 11, in <module>
    from logdetective.server.models import AnalyzedSnippet
ModuleNotFoundError: No module named 'logdetective.server.models'
